rename reducers, remove  warnings
