+++ src/components/Search/Search.js
@@ -5,7 +5,8 @@ const Search = (props) => {
   return (
     <div className='searchBox'>
       <input type="text" name="search" placeholder="Search" onChange={e => {
-        props.handleSearch(e.target.value);}
+        props.handleSearch(e.target.value);
+      }
       }/>
     </div>
   );
