+++ src/index.js
@@ -1,9 +1,9 @@
 require('./scss/index.scss');
 import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
-import { BrowserRouter as Router, Route, Switch} from 'react-router-dom';
-import {createStore, combineReducers, compose} from 'redux';
-import {Provider} from 'react-redux';
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import { createStore, combineReducers, compose } from 'redux';
+import { Provider } from 'react-redux';
 import rootReducer from './reducer/reducer';
 import DeviceList from './pages/DeviceList/DeviceList';
 import DevicePage from './pages/DevicePage/DevicePage';
